Extract Unix Python stdin protocol
